+++ habit/src/main/java/app/habit/service/AnswerService.java
@@ -1,4 +1,4 @@
-package app.habit.service.gpt;
+package app.habit.service;
 
 import app.habit.domain.Answer;
 import app.habit.domain.HabitAssessmentManager;
