+++ habit/src/main/java/app/habit/service/FeedbackSessionService.java
@@ -1,4 +1,4 @@
-package app.habit.service.gpt;
+package app.habit.service;
 
 import app.habit.domain.FeedbackSession;
 import app.habit.repository.FeedbackSessionRepository;
