+++ habit/src/main/java/app/habit/service/HabitFormingPhaseService.java
@@ -1,4 +1,4 @@
-package app.habit.service.gpt;
+package app.habit.service;
 
 import app.habit.domain.HabitFormingPhase;
 import app.habit.repository.HabitFormingPhaseRepository;
